extracted file card component into an independent module
