+++ app/frontend/src/components/Dropzone.tsx
@@ -1,11 +1,12 @@
-import { cn, computeFileSize } from "@/lib/utils";
+import { cn } from "@/lib/utils";
 import { useDropzone } from "react-dropzone";
-import { Loader, UploadIcon } from "lucide-react";
+import { UploadIcon } from "lucide-react";
 import { useCallback, useState } from "react";
 import { Button } from "@/components/ui/button";
+import FileCard from "@/components/FileCard";
 
 type APIResponse = {
-  payload: string;
+  payload: string[];
   filename: string;
   pages: string;
 };
@@ -16,7 +17,7 @@ export default function Dropzone(props: React.HTMLAttributes<HTMLDivElement>) {
   const [internalError, setInternalError] = useState<string | null>(null);
   const [uploadError, setUploadError] = useState<string | null>(null);
   const [isUploading, setIsUploading] = useState(false);
-  const [payload, setPayload] = useState<APIResponse | null>(null);
+  const [result, setResult] = useState<APIResponse | null>(null);
   const totalFileSize = file?.size ?? 0;
 
   const onDrop = useCallback((acceptedFiles: File[]) => {
@@ -66,7 +67,7 @@ export default function Dropzone(props: React.HTMLAttributes<HTMLDivElement>) {
       );
       const result = await response.json();
 
-      setPayload(result as APIResponse);
+      setResult(result as APIResponse);
     } catch (error) {
       console.error(error);
     } finally {
@@ -115,7 +116,7 @@ export default function Dropzone(props: React.HTMLAttributes<HTMLDivElement>) {
   };
 
   // initial render
-  if (!payload) {
+  if (!result) {
     return file ? (
       <div className="space-y-4">
         <FileCard
@@ -138,67 +139,10 @@ export default function Dropzone(props: React.HTMLAttributes<HTMLDivElement>) {
   return (
     <div className="space-y-4">
       <div className="space-y-2 lg:space-y-0 lg:grid lg:grid-cols-2 lg:gap-3">
-        <p>{payload["payload"]}</p>
+        <p>{result["payload"]}</p>
         {/* <ResultCard title={title} thumbUrl={thumb_url} url={url} /> */}
       </div>
-      <Button onClick={() => setPayload(null)}>Clear</Button>
-    </div>
-  );
-}
-
-function FileCard({
-  file,
-  setFile: setFiles,
-  isUploading,
-  handleUpload,
-  setUploadError,
-}: {
-  file: File;
-  setFile: React.Dispatch<React.SetStateAction<File | null>>;
-  setUploadError: React.Dispatch<React.SetStateAction<string | null>>;
-  isUploading: boolean;
-  handleUpload: (file: File) => void;
-}) {
-  const removeFile = () => {
-    setUploadError(null);
-    setFiles(null);
-  };
-
-  return (
-    <div className="mx-auto max-w-[700px] flex items-center justify-between p-3 bg-white rounded-md border border-gray-200 shadow">
-      <div className="flex items-center gap-2">
-        <div className="w-16 h-16 bg-gray-300 rounded flex items-center justify-center p-5">
-          <span className="text-xs font-lg">
-            {file.name.split(".").pop()?.toUpperCase()}
-          </span>
-        </div>
-
-        <div className="flex flex-col items-start space-y-1">
-          <p className="text-sm font-medium truncate max-w-xs sm:max-w-[30rem]">
-            {file.name}
-          </p>
-          <p className="text-xs text-gray-500">{computeFileSize(file)}</p>
-        </div>
-      </div>
-
-      <div className="flex space-x-2">
-        <Button
-          onClick={() => handleUpload(file)}
-          className="cursor-pointer"
-          disabled={isUploading}
-        >
-          {isUploading && <Loader className="animate-spin" />}
-          {isUploading ? "Uploading" : "Upload"}
-        </Button>
-
-        <Button
-          variant="destructive"
-          className="cursor-pointer"
-          onClick={() => removeFile()}
-        >
-          Remove
-        </Button>
-      </div>
+      <Button onClick={() => setResult(null)}>Clear</Button>
     </div>
   );
 }
